refactor scripts and create remove scripts for each process

+++ setup_conda_env.bat
@@ -1,15 +1,26 @@
-set root=C:\ProgramData\Anaconda3
-set newenv=hypergan_test
-
 echo OFF
+
+set root=C:\ProgramData\Anaconda3
+set targetenv=hypergan_test
 
 call %root%\Scripts\activate.bat %root%
 
 call conda update conda
-call conda env remove --name %newenv%
-call conda create -n %newenv% python=3.6 anaconda
-call conda activate %newenv%
-call pip install tensorflow-gpu hyperchamber nashpy pygame natsort opencv-python  
+call conda activate %targetenv%
+
+set envpath=%CONDA_PREFIX%
+for %%F in (%envpath%) do set thisenv=%%~nxF
+
+if %thisenv% == %targetenv% (
+	echo Target conda env, %targetenv%, already exists. Skipping conda env creation
+) else (
+	echo Creating conda environment %targetenv%
+	call conda create -n %targetenv% python=3.6 anaconda
+	call conda activate %targetenv%
+	call pip install tensorflow-gpu hyperchamber nashpy pygame natsort opencv-python  
+)
+
+
