+++ setup_fashion_mnist.cmd
@@ -1,24 +1,40 @@
-set root=C:\ProgramData\Anaconda3
-set newenv=hypergan_test
-set hypergan=https://github.com/HyperGAN/HyperGAN.git
-set hgfolder=HyperGAN
-set traincommit=develop
-set samplecommit=develop
-set fashionfile=train-images-idx3-ubyte.gz
-set fashionbase=http://fashion-mnist.s3-website.eu-central-1.amazonaws.com/
-set fashionurl=%fashionbase%%fashionfile%
-set fashiondir=fashion_mnist
-set filetype=png
-
 echo OFF
 
-del %fashionfile%
-curl -O %fashionurl%
+set root=C:\ProgramData\Anaconda3
+set targetenv=hypergan_test
+set fashionfile=train-images-idx3-ubyte
+set gz=.gz
+set fashionfilegz=%fashionfile%%gz%
+set fashionbase=http://fashion-mnist.s3-website.eu-central-1.amazonaws.com/
+set fashionurl=%fashionbase%%fashionfilegz%
+set fashiondir=fashion_mnist
+
+IF  EXIST %fashionfilegz% (
+	echo %fashionfilegz% EXISTS. Skipping downloading dataset.
+) else (
+	echo %fashionfilegz% doesn't exist. Downloading.
+	curl -O %fashionurl%
+)
 
 call %root%\Scripts\activate.bat %root%
-call conda activate %newenv%
+call conda activate %targetenv%
+set envpath=%CONDA_PREFIX%
+for %%F in (%envpath%) do set thisenv=%%~nxF
 
-call python convert_fashion_mnist.py
+If EXIST %fashionfile% (
+	echo %fashionfile% EXISTS. Skipping unzipping.
+) else (
+	echo %fashionfile% doesn't exist. Unzipping.
+	
+	if "%thisenv%" == "%targetenv%" (
+		echo Unzipping fashion mnist dataset and extracting images.
+		call python convert_fashion_mnist.py
+	) else (
+		echo Unable to activate %targetenv% conda environment. %envpath% / %thisenv%
+		echo Program will now end. Run setup_conda_env.bat first and verify environment is properly installed.
+		exit
+	)	
+)
